object: drop Range API support, fix #4136
Leave stub at the node level, drop everything related otherwise, it's
deprecated since API 2.23 and no longer required for applications.
