+++ cmd/neofs-cli/modules/acl/extended/create.go
@@ -24,7 +24,7 @@ Rule consist of these blocks: <action> <operation> [<filter1> ...] [<target1> ..
 
 Action is 'allow' or 'deny'.
 
-Operation is an object service verb: 'get', 'head', 'put', 'search', 'delete', or 'getrange'.
+Operation is an object service verb: 'get', 'head', 'put', 'search' or 'delete'.
 
 Filter consists of <typ>:<key><match><value>
   Typ is 'obj' for object applied filter or 'req' for request applied filter.
