+++ cmd/neofs-cli/modules/object/get.go
@@ -1,6 +1,8 @@
 package object
 
 import (
+	"bytes"
+	"errors"
 	"fmt"
 	"io"
 	"os"
@@ -21,6 +23,10 @@ import (
 )
 
 const (
+	rangeFlag      = "range"
+	rangeFlagUsage = "Range to take data from in the form offset:length"
+	rangeSep       = ":"
+
 	payloadOnlyFlag   = "payload-only"
 	extendedRangeFlag = "extended-range"
 )
@@ -268,4 +274,92 @@ func payloadReadSize(payloadSize uint64, ranges []*object.Range, first, last *ui
 	default:
 		return int64(payloadSize)
 	}
+}
+
+func printSplitInfoErr(cmd *cobra.Command, err error) (bool, error) {
+	errSplitInfo, ok := errors.AsType[*object.SplitInfoError](err)
+
+	if ok {
+		cmd.PrintErrln("Object is complex, split information received.")
+		if err := printSplitInfo(cmd, errSplitInfo.SplitInfo()); err != nil {
+			return false, err
+		}
+	}
+
+	return ok, nil
+}
+
+func printSplitInfo(cmd *cobra.Command, info *object.SplitInfo) error {
+	bs, err := marshalSplitInfo(cmd, info)
+	if err != nil {
+		return fmt.Errorf("can't marshal split info: %w", err)
+	}
+
+	cmd.Println(string(bs))
+	return nil
+}
+
+func marshalSplitInfo(cmd *cobra.Command, info *object.SplitInfo) ([]byte, error) {
+	toJSON, _ := cmd.Flags().GetBool(commonflags.JSON)
+	toProto, _ := cmd.Flags().GetBool("proto")
+	switch {
+	case toJSON && toProto:
+		return nil, errors.New("'--json' and '--proto' flags are mutually exclusive")
+	case toJSON:
+		return info.MarshalJSON()
+	case toProto:
+		return info.Marshal(), nil
+	default:
+		b := bytes.NewBuffer(nil)
+		if splitID := info.SplitID(); splitID != nil {
+			b.WriteString("Split ID: " + splitID.String() + "\n")
+		}
+		if link := info.GetLink(); !link.IsZero() {
+			b.WriteString("Linking object: " + link.String() + "\n")
+		}
+		if first := info.GetFirstPart(); !first.IsZero() {
+			b.WriteString("First object: " + first.String() + "\n")
+		}
+		if last := info.GetLastPart(); !last.IsZero() {
+			b.WriteString("Last object: " + last.String() + "\n")
+		}
+		return b.Bytes(), nil
+	}
+}
+
+func getRangeList(cmd *cobra.Command) ([]*object.Range, error) {
+	v := cmd.Flag("range").Value.String()
+	if len(v) == 0 {
+		return nil, nil
+	}
+	vs := strings.Split(v, ",")
+	rs := make([]*object.Range, len(vs))
+	for i := range vs {
+		offString, lenString, found := strings.Cut(vs[i], rangeSep)
+		if !found {
+			return nil, fmt.Errorf("invalid range specifier: %s", vs[i])
+		}
+
+		offset, err := strconv.ParseUint(offString, 10, 64)
+		if err != nil {
+			return nil, fmt.Errorf("invalid '%s' range offset specifier: %w", vs[i], err)
+		}
+		length, err := strconv.ParseUint(lenString, 10, 64)
+		if err != nil {
+			return nil, fmt.Errorf("invalid '%s' range length specifier: %w", vs[i], err)
+		}
+
+		if length == 0 {
+			if offset != 0 {
+				return nil, fmt.Errorf("invalid '%s' range: zero length with non-zero offset", vs[i])
+			}
+		} else if offset+length <= offset {
+			return nil, fmt.Errorf("invalid '%s' range: uint64 overflow", vs[i])
+		}
+
+		rs[i] = object.NewRange()
+		rs[i].SetOffset(offset)
+		rs[i].SetLength(length)
+	}
+	return rs, nil
 }
