+++ cmd/neofs-cli/modules/util/acl.go
@@ -20,11 +20,10 @@ import (
 func PrettyPrintTableBACL(cmd *cobra.Command, bacl *acl.Basic) {
 	// Header
 	w := tabwriter.NewWriter(cmd.OutOrStdout(), 1, 4, 4, ' ', 0)
-	fmt.Fprintln(w, "\tRange\tSearch\tDelete\tPut\tHead\tGet")
+	fmt.Fprintln(w, "\tSearch\tDelete\tPut\tHead\tGet")
 	// Bits
 	bits := []string{
 		boolToString(bacl.Sticky()) + " " + boolToString(!bacl.Extendable()),
-		getRoleBitsForOperation(bacl, acl.OpObjectRange),
 		getRoleBitsForOperation(bacl, acl.OpObjectSearch), getRoleBitsForOperation(bacl, acl.OpObjectDelete),
 		getRoleBitsForOperation(bacl, acl.OpObjectPut), getRoleBitsForOperation(bacl, acl.OpObjectHead),
 		getRoleBitsForOperation(bacl, acl.OpObjectGet),
@@ -32,7 +31,7 @@ func PrettyPrintTableBACL(cmd *cobra.Command, bacl *acl.Basic) {
 	fmt.Fprintln(w, strings.Join(bits, "\t"))
 	// Footer
 	footer := []string{"X F"}
-	for range 6 {
+	for range 5 {
 		footer = append(footer, "U S O B")
 	}
 	fmt.Fprintln(w, strings.Join(footer, "\t"))
